# Improve cartodb._CDB_total_relation_size performance