@rochoa 2015-08-17
BTW: 
> **Tip**: A block containing an EXCEPTION clause is significantly more expensive to enter and exit than a block without one. Therefore, don't use EXCEPTION without need.

From [PL/pgSQL - SQL Procedural Language: 40.6. Control Structures](http://www.postgresql.org/docs/9.3/static/plpgsql-control-structures.html).