@javisantana 2015-08-17
ok. I think that maybe the solution for this is not to use CTE in the check quota query